+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/eventpublishing/EventLinkingTimerBean.java
@@ -1,12 +1,14 @@
 package uk.gov.justice.services.eventsourcing.eventpublishing;
 
 import uk.gov.justice.services.ejb.timer.TimerServiceManager;
+import uk.gov.justice.services.eventsourcing.eventpublishing.configuration.EventLinkingWorkerConfig;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
 import javax.ejb.Singleton;
 import javax.ejb.Startup;
 import javax.ejb.Timeout;
+import javax.ejb.Timer;
 import javax.ejb.TimerService;
 import javax.inject.Inject;
 
@@ -14,13 +16,13 @@ import javax.inject.Inject;
 @Startup
 public class EventLinkingTimerBean {
 
-    static final String TIMER_JOB_NAME = "event-store.link-events.worker";
+    static final String TIMER_JOB_NAME = "event-store.link-new-events.job";
 
     @Resource
     private TimerService timerService;
 
     @Inject
-    private EventPublishingWorkerTimerConfig eventPublishingWorkerTimerConfig;
+    private EventLinkingWorkerConfig eventLinkingWorkerConfig;
 
     @Inject
     private TimerServiceManager timerServiceManager;
@@ -28,18 +30,25 @@ public class EventLinkingTimerBean {
     @Inject
     private EventLinkingWorker eventLinkingWorker;
 
+    @Inject
+    private SufficientTimeRemainingCalculatorFactory sufficientTimeRemainingCalculatorFactory;
+
     @PostConstruct
     public void startTimerService() {
-
         timerServiceManager.createIntervalTimer(
                 TIMER_JOB_NAME,
-                eventPublishingWorkerTimerConfig.getTimerStartWaitMilliseconds(),
-                eventPublishingWorkerTimerConfig.getTimerIntervalMilliseconds(),
+                eventLinkingWorkerConfig.getTimerStartWaitMilliseconds(),
+                eventLinkingWorkerConfig.getTimerIntervalMilliseconds(),
                 timerService);
     }
 
     @Timeout
-    public void runEventLinkingWorker() {
-        eventLinkingWorker.findAndLinkUnlinkedEvents();
+    public void runEventLinkingWorker(final Timer timer) {
+
+        final SufficientTimeRemainingCalculator sufficientTimeRemainingCalculator = sufficientTimeRemainingCalculatorFactory.createNew(
+                timer,
+                eventLinkingWorkerConfig.getTimeBetweenRunsMilliseconds());
+
+        eventLinkingWorker.linkNewEvents(sufficientTimeRemainingCalculator);
     }
 }