+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/eventpublishing/LinkedEventPublisher.java
@@ -28,7 +28,7 @@ public class LinkedEventPublisher {
     private EventPublishingRepository eventPublishingRepository;
 
     @Transactional(REQUIRES_NEW)
-    public boolean publishNextQueuedEvent() {
+    public boolean publishNextNewEvent() {
 
         final Optional<UUID> eventId = eventPublishingRepository.getNextEventIdFromPublishQueue();
         if (eventId.isPresent()) {